Revert "Don't trigger TLS handshake until first I/O operation."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/53794c7f-76dd-4265-b2ae-b2e31f24e340). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic6754528a9409acf4769d753288f4268949330ee