Revert "Don't trigger TLS handshake until first I/O operation."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a0db35bc-44a1-4450-bdff-eaa42cfd0808). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8ea748e07071d08fc7639b7f8a9de6e83b0c952b