Revert "Don't trigger TLS handshake until first I/O operation."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4cce7496-0235-451c-a0de-3e91829630fd). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id20fa061db6ec3024509f38edd21d53f537a2595