Merge: Make tests agnostic about TLS v1.x.
Merged from upstream commit #1150.

Test only change. Ensures tests neither use nor assume anything
about whether TLSv.1 are enabled or supported.  As such it is
suitable for backporting to historic Android test suites where
vendors may have disabled TLS v1.x by editing the default arrays.

However by being agnostic it does not enforce that TLS v1.x are
available if expected. A further change will provide an API for
that, but which is not suitable for backporting as it will
require non-test changes.

A lot of the tidy-up is around RenegotiationTest and its
TestUtils methods, which are only tested on OpenJDK builds.

Bug: 288058920
Bug: 288062754
Test: atest MtsLibcoreTestCases (with and without TLS v1.x enabled)
Change-Id: Ic22c4a435d8651d37cc31bafa9c74accc1f690ca