Revert "Update apex/tests/AndroidManifest.xml"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2a1f7796-ee58-45fc-a18e-739cddb682a5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2ab3b581ea4cfbf50d187082625cf4ea9a844fd8