Revert "Remove "extends TestCase" from TrustedCertificateStoreTest."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/dbd9e3f0-3b64-4ee2-9351-dc027081bacf). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2acabfcf1239704a0b9993a0690c350b96522ab6