Reset HMAC and CMAC contexts more efficiently
Cherry-picked from upsteram PR #1166.

Part of a series to undo the collateral damage of
https://r.android.com/2821976 which reverted a number of
innocent PRs from https://r.android.com/2810393 along with
the intended one.

If computing a new MAC with the same key, we would ideally save a little
bit of work. (A couple blocks of the underlying hash for HMAC and key
schedule calculation with CMAC.) Use the APIs for doing this, and also
cover this case with tests.

I noticed that MacSpi also has an optional clone() method, which we
could support here too, but I've left that for another PR.

Test: MTS
Change-Id: I331dad911716d1c843e4a8870f6a1e25e2738a5b