DO NOT SUBMIT: Avoid static-initializer recursion
Do not submit this, this is a change in generated code.

What appears to be happening is that when 'setDefaultUserDirectory' is
called for the first time, the inner class PreloadHolder has not yet
been loaded by the VM. Setting the value of a variable in PreloadHolder
triggers it's static initializer, which calls 'setDefaultUserDirectory'.

My unfounded theory: our VM does not seem to be happy when it enters the
'setDefaultUserDirectory' function again, because when it goes to set
the variable on PreloadHolder, it *may* be upset that the static
initializer has not finished. This causes the program to hang.

As a quick band-aid, we can move the implementation of this function
into the inner class and eliminate calls that originate from the inner
class to the outer class.

Bug: 305970538
Change-Id: Ia2dca96071dfd6fe8d4db809498bb7da1c574bc2