Don't throw when resetting an uninitialised Mac. (#1178)
Cherry-picked from upstream PR #1178 and needed because
https://r.android.com/2810393 inadvertantly changed this
behaviour.

Turns out Android, SunJCE and BC all allow this and the javadoc
has nothing to say on the subject.  Caught this as a side
effect of an unrelated Android test, so added an explicit check
in our own ServiceTester test which runs against all installed
Providers.

Added a new field to track initialized state because if the
native context ever becomes null after that (e.g. concurrency
bug) that's still an error.

Test: MtsConscryptTestCases:org.apache.harmony.crypto.tests.javax.crypto.MacTest
Change-Id: Iba1c8ad2653b9b190879b336fc51c1b544d92a1c