Add Conscrypt "Lite" client library.
This jarjars a new target from the (very) small subset of
Conscrypt needed to access its pseudo-JCA APIs such as HPKE
which aren't available via public APIs (yet).

The resulting library is about 20KB and can be used in both
host and device builds.  It will mostly work as-is with
Mainline Conscrypt, but functionality such as raw key specs
and interoperability with the RI Xec key secs will not work
until https://r.android.com/2887887 lands in Mainline.

Bug: 300089422
Test: vogar `cparg conscrypt-lite`
Change-Id: I870ffaac972d40d68b84fa2a3445979163d265f1