Revert "Use srcs attribute for prebuilt_etc"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/90ce5c2f-be6c-4d97-bae1-3dee6fc43912). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8c2fe36c6174bb019899a7f020f44dfcff63a8df