Revert "Minor Scrypt tidy-ups."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e2efc57c-4c5a-44df-8cf7-d6f0f8f4ef35). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Icd6340e3c9ddf3051b95e37066ce03ad7071735a