Revert "Add scrypt support."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/40ff83e2-140d-48e5-817c-818056ca964b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Icb3891d1b2438cbd0d0763c8dd487208db277307