Revert "Remove InternalUtil"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/419d95a7-2f8a-4ced-a1b1-10f613c0ab96). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0415c41bc343f6ef50e94e707b58f78f2dbdeec6
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com