Revert "Remove InternalUtil"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9c9f3325-c94d-4fd7-b5ad-e5278689da1a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I41173b5be17b883bc06eb091b95793de3556db19
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com