Revert "Remove ct fallback logs"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c611145d-f79f-4dac-be3f-e78343522776). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 350034266
Change-Id: If3ce051731fa6371779304d9dd9fc88ecbff0fe1
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com