Revert "Enable CertBlocklistFileTest in TEST_MAPPING"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9257f983-98ab-462f-8b53-8f33d011d7bf). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3283a7d36151b63bda80fe8e05ce4531f3c3709d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com