Revert "Disable from-text stub generation for conscrypt.module.i..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5577acbc-0e94-4f22-b3e1-a52b1b440ffe). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 357656342
Change-Id: I10aa7d7d5832c02bd3952a8a672c7c3fe9fa578f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com