Revert "Disable from-text stub generation for conscrypt.module.i..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1bdf2467-c0e3-472b-a704-bc4c98cb6dbb). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0b3a87c968fb6747d5f3079e03f47365d3ce180d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com