Filter out SCTs emitted after a log expired
This is a requirement from the policy (see filterRetiredExpired).

The state of a LogInfo is associated with a timestamp, populated from
the log list.

Timestamps are stored internally as long. The date parsing (ISO 8601) is
done while processing the log list. This can be later optimized by
changing the log list timestamps to be Unix times.

SimpleDateFormat is used as other more recent alternatives
(i.e., DateTimeFormatter) are only available in API 26+.

Bug: 319829948
Test: atest
Change-Id: Icf2fb304f03c6ededba0ab8931ee294a86b05229