WIP: starting work for conscrypt cipher metrics
We should of course properly create the atom before submitting this but wanted to do a quick outline of how the metric logging would happen. In addition to this the next step would also be to add a mechanism that combines results and logs them at certain intervals instead of logging each atom as it comes.

Change-Id: I1846cee1ce74b17469b96cbb664174ca5d749f1b