Revert "Filter out SCTs emitted after a log expired"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e476cd7c-d853-4de7-8696-14a945d2bb34). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I49c90df6d169b06c6ae0a24cd9471b6b4370a6be
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com