Add metric reporting for log list status
For the Android platform, report to statsd when a log list has been
loaded (successfully or not).

A new interface StatsLog is added to abstract ConscryptStatsLog (which
is renamed to StatsLogImpl). This interface is used for dependency
injection in LogStoreImpl.

The Platform class is marked as public so that the LogStoreImpl (which
belongs to the ct subpackage) can access it.

The metrics methods (countTlsHandshake and the new
updateCTLogListStatusChanged) are moved out of Platform into the new
StatsLog. Platform is updated to simply return the implementation
instance (singleton).

Test: adb shell cmd stats print-logs; run ctdemo manually; confirm that
      the atom is sent to statsd correctly (via logcat)
Test: atest
Bug: 356111598
Bug: 319829948
Change-Id: I9f2709ef6c1acf349d197560c7580e8bf43c5281