Revert "Remove CT tests"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fb0f5164-b67a-48d6-b1e7-a8d95dea20a7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 369929601
Change-Id: Ia9d67e925a4c2d338c95f7d420c9cd7c846b569a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com