Revert "Extract test file lists out to filegroups."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8d18e634-7ea1-4a87-9e9f-355d1b02b1ff). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 376106028
Change-Id: Iedc11d8f9cf806c36c196f61510023048ebc2244
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com