Revert "Implement Android CT Policy for OCSP and TLS Data SCTs"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6d1ed5a0-8da3-41a3-bb8a-588ed76942ad). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I798cb5545235afc33ae8a72674a1537ebc026a98
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com