Decouple TrustManagerImpl from CT classes
TrustManagerImpl had multiple references to Certificate Transparency
objects: LogStore, Verifier, Policy, etc. Abstract these into a
unique class (ct.SubSsytem).

One constructor for TrustManagerImpl (4 parameters) is removed as it is
not used and not part of the API. The largest constructor is updated to
simply receive an ct.SubSystem instance.

This change is a no-op. It will help testing the logic of checkCT.

Bug: 356111598
Test: atest CtsLibcoreTestCases
Test: atest ConscryptPrivateTestCases
Change-Id: I03c5fd94aa5912035b54bb0912167f9e421de33b