Remove LogStore.setPolicy
LogStore requires access to the policy when loading, to determine if the
log list is considered compliant or not. The policy was previously
associated using a separate setter (setPolicy). Force the passing of the
policy to the constructor of LogStore to avoid potentially missing that
call.

This is a no-op change.

Bug: 356111598
Test: atest ConscryptPrivateTestCases
Change-Id: Icfb5e3673d357abb59073509b976c5be342bdf08