Refactor metrics classes
Regenerate ConscryptStatsLog based on the latest proto definition. This
class is nearly identical to the generated wrapper except for the usage
of ReflexiveStatsEvent and ReflexiveStatsLog. Updated the class
documentation with the constraints around its usage.

buildEvent methods are not used as non-existent in the wrapper. They are
marked @Deprecated but not removed because implicitly used by
MetricsTest in CTS. ReflexiveStatsEvent builder is updated to ignore any
call to writeIntArray if not supported by the platform (i.e., sdk < 32).

A new class (Constants) is added to expose the generated constants in a
more succinct way to the rest of Conscrypt. The Source class is removed
as redundant with Constants (this also ensures that the Source values
remain in sync with the proto).

From a behaviour's perspective, this change is a no-op.

Test: atest
Bug: 319829948
Change-Id: Ie6b166530544345491347ff6bda8bbf4633e37c3