Revert "Rework the SPAKE2+ API"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3fbaddce-3d86-469e-937e-34ab4d070679). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ifbc60f1ebdb5ba5bb332fa5a92f8ebc0bf404bf8
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com