Revert "Fix validation logic for SPAKE2PLUS_PRERELEASE."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1e076a9d-f9bc-43ec-9f8e-157f8818aa90). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I32c5558eb418685efee0cc5891e409d1c8ac0989
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com